+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Checkout/Checkout+Session+DebugDescription.swift
@@ -96,7 +96,7 @@ extension CheckoutController.Session: CustomDebugStringConvertible {
         if let billingDetails = paymentOption.billingDetails {
             lines.append(contentsOf: addressDebugDescriptionLines(
                 name: "billingDetails",
-                country: billingDetails.address.country,
+                country: billingDetails.address?.country,
                 indentation: 2
             ))
         } else {